feat: extract topology builder, preserve sessionId on agent complete
- Extract shared buildTopologyFromEvents() from live-task-view into
  src/lib/topology/build-from-events.ts and reuse in both LiveTaskView
  and ContainerAgentPanel TopologyTab
- Preserve sessionId on task when agent completes/errors/cancels so the
  UI can display session events and topology for review before approval
- Add E2E topology screenshots
